requirements.txt
pygame==2.6.1
pytest==8.3.3
windows-curses
Uni-Curses==3.1.2
Jinja2~=3.1.4
pillow~=11.1.0